Fix an overflow in calculating trigger current values.
This fixes the BTS7960 5A/10A motor board calculations, the ESP32
appears to have been overflowing during the dynamic calculations. For
the Arduino motor board with only a 2A limit the calculations were
coming out correctly but with the BTS7960 having a 43A limit they were
overflowing.

+++ src/MotorBoard.h
@@ -60,8 +60,8 @@ private:
 	const adc1_channel_t _senseChannel;
 	const uint8_t _enablePin;
 	const uint32_t _maxMilliAmps;
-	const uint16_t _triggerValue;
-	uint16_t _current;
+	const uint32_t _triggerValue;
+	uint32_t _current;
 	uint32_t _lastCheckTime;
 	bool _state;
 	bool _triggered;
